Rename option package to config
