+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ## v2.0.0 (2024-10-18)
 ### Breaking changes
-- ✨ Added the ability to register your own custom functions for specific types and use them in your Textwire code like built-in functions
+- ✨ Added the ability to register your own custom functions for specific types and use them in your Textwire code like built-in functions. If you are upgrading from version 1, change `textwire.Config`
 
 ## v1.7.1 (2024-09-08)
 - 🧑‍💻 Improve error handling for component slots. When you pass a slot that isn't defined, you'll get an error
