+++ parsing.go
@@ -4,9 +4,9 @@ import (
 	"strings"
 
 	"github.com/textwire/textwire/v2/ast"
+	"github.com/textwire/textwire/v2/config"
 	"github.com/textwire/textwire/v2/fail"
 	"github.com/textwire/textwire/v2/lexer"
-	"github.com/textwire/textwire/v2/option"
 	"github.com/textwire/textwire/v2/parser"
 )
 
@@ -120,7 +120,7 @@ func applyComponentToProgram(prog *ast.Program, progFilePath string) *fail.Error
 	return nil
 }
 
-func applyOptions(opt *option.Option) {
+func applyOptions(opt *config.Config) {
 	usesTemplates = true
 
 	if opt.TemplateDir != "" {
